move test and pass content to getWelcomeTarget
