+++ jetty-core/jetty-server/src/main/java/org/eclipse/jetty/server/ResourceService.java
@@ -460,7 +460,7 @@ public class ResourceService
         }
 
         // process optional Welcome behaviors
-        if (welcome(request, response, callback))
+        if (welcome(content, request, response, callback))
             return;
 
         if (!passConditionalHeaders(request, response, content, callback))
@@ -499,9 +499,9 @@ public class ResourceService
     {
     }
 
-    private boolean welcome(Request request, Response response, Callback callback) throws Exception
+    private boolean welcome(HttpContent content, Request request, Response response, Callback callback) throws Exception
     {
-        WelcomeAction welcomeAction = processWelcome(request);
+        WelcomeAction welcomeAction = processWelcome(content, request);
         if (LOG.isDebugEnabled())
             LOG.debug("welcome(req={}, rsp={}, cbk={}) welcomeAction={}", request, response, callback, welcomeAction);
 
@@ -581,9 +581,9 @@ public class ResourceService
         Response.writeError(request, response, callback, HttpStatus.INTERNAL_SERVER_ERROR_500);
     }
 
-    private WelcomeAction processWelcome(Request request) throws IOException
+    private WelcomeAction processWelcome(HttpContent content, Request request) throws IOException
     {
-        String welcomeTarget = getWelcomeFactory().getWelcomeTarget(request);
+        String welcomeTarget = getWelcomeFactory().getWelcomeTarget(content, request);
         if (welcomeTarget == null)
             return null;
 
@@ -892,7 +892,7 @@ public class ResourceService
          * @return The URI path of the matching welcome target in context or null
          * if no welcome target was found
          */
-        String getWelcomeTarget(Request request) throws IOException;
+        String getWelcomeTarget(HttpContent content, Request request) throws IOException;
     }
 
     private static class ContentWriterIteratingCallback extends IteratingCallback
