+++ jetty-ee10/jetty-ee10-servlet/src/test/java/org/eclipse/jetty/ee10/servlet/DefaultServletTest.java
@@ -3185,6 +3185,63 @@ public class DefaultServletTest
         }
     }
 
+    @Test
+    public void testPathInfoOnly() throws Exception
+    {
+        ServletContextHandler context = new ServletContextHandler(null, "/c1", ServletContextHandler.NO_SESSIONS);
+        context.setWelcomeFiles(new String[]{"index.y", "index.x"});
+        ServletHolder indexServlet = new ServletHolder("index-servlet", new HttpServlet()
+        {
+            @Override
+            protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException
+            {
+                resp.setContentType("text/plain");
+                resp.setCharacterEncoding("UTF-8");
+                resp.getWriter().println(req.getRequestURI());
+                resp.getWriter().println("testPathInfoOnly-OK");
+                resp.getWriter().close();
+            }
+        });
+
+        ServletMapping indexMapping = new ServletMapping();
+        indexMapping.setServletName("index-servlet");
+        indexMapping.setPathSpecs(new String[]{"*.x", "*.y"});
+        context.getServletHandler().addServlet(indexServlet);
+        context.getServletHandler().addServletMapping(indexMapping);
+
+        Path pathTest = MavenTestingUtils.getTestResourcePath("pathTest");
+
+        Path defaultDir = pathTest.resolve("default");
+        ServletHolder slashHolder = new ServletHolder("default", new DefaultServlet());
+        slashHolder.setInitParameter("redirectWelcome", "false");
+        slashHolder.setInitParameter("welcomeServlets", "true");
+        slashHolder.setInitParameter("pathInfoOnly", "false");
+        slashHolder.setInitParameter("baseResource", defaultDir.toAbsolutePath().toString());
+        context.addServlet(slashHolder, "/");
+
+        Path rDir = pathTest.resolve("rdir");
+        ServletHolder rHolder = new ServletHolder("rdefault", new DefaultServlet());
+        rHolder.setInitParameter("redirectWelcome", "false");
+        rHolder.setInitParameter("welcomeServlets", "true");
+        rHolder.setInitParameter("pathInfoOnly", "true");
+        rHolder.setInitParameter("baseResource", rDir.toAbsolutePath().toString());
+        context.addServlet(rHolder, "/r/*");
+
+        server.stop();
+        server.setHandler(context);
+        server.start();
+        String rawRequest = """
+            GET /c1/r/ HTTP/1.1\r
+            Host: localhost\r
+            Connection: close\r
+            \r
+            """;
+
+        String rawResponse = connector.getResponse(rawRequest);
+        HttpTester.Response response = HttpTester.parseResponse(rawResponse);
+        assertThat(response.getContent(), containsString("testPathInfoOnly-OK"));
+    }
+
     public static class WriterFilter implements Filter
     {
         @Override
